move tests out of document and update diff test.
