+++ identity_core/src/document.rs
@@ -93,74 +93,3 @@ impl FromStr for DIDDocument {
         Ok(doc)
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use serde_diff::{Apply, Diff};
-
-    /// test doc creation via the `DIDDocument::new` method.
-    #[test]
-    fn test_doc_creation() {
-        let mut did_doc =
-            DIDDocument::new("https://w3id.org/did/v1".into(), "did:iota:123456789abcdefghi".into()).unwrap();
-        let service = Service::new(
-            "did:into:123#edv".into(),
-            "EncryptedDataVault".into(),
-            "https://edv.example.com/".into(),
-            None,
-            None,
-        )
-        .unwrap();
-        did_doc.add_service(service.clone());
-        let public_key = PublicKey::new(
-            "did:iota:123456789abcdefghi#keys-1".into(),
-            "RsaVerificationKey2018".into(),
-            "did:iota:123456789abcdefghi".into(),
-            "publicKeyBase58".into(),
-            "H3C2AVvLMv6gmMNam3uVAjZpfkcJCwDwnZn6z3wXmqPV".into(),
-        )
-        .unwrap();
-        did_doc.add_key_pair(public_key.clone());
-
-        let mut did_doc_2 =
-            DIDDocument::new("https://w3id.org/did/v1".into(), "did:iota:123456789abcdefghi".into()).unwrap();
-        did_doc_2.add_service(service);
-        did_doc_2.add_key_pair(public_key);
-
-        let did_doc = did_doc.init_timestamps().unwrap();
-
-        // did_doc has timestamps while did_doc_2 does not.
-        assert_ne!(did_doc, did_doc_2);
-    }
-
-    #[test]
-    fn test_diff() {
-        let mut old = DIDDocument::new("https://w3id.org/did/v1".into(), "did:iota:123456789abcdefghi".into()).unwrap();
-        let mut new = DIDDocument::new("https://w3id.org/did/v1".into(), "did:iota:123456789abcdefghi".into()).unwrap();
-        let service = Service::new(
-            "did:into:123#edv".into(),
-            "EncryptedDataVault".into(),
-            "https://edv.example.com/".into(),
-            None,
-            None,
-        )
-        .unwrap();
-        new.add_service(service.clone());
-        let public_key = PublicKey::new(
-            "did:iota:123456789abcdefghi#keys-1".into(),
-            "RsaVerificationKey2018".into(),
-            "did:iota:123456789abcdefghi".into(),
-            "publicKeyBase58".into(),
-            "H3C2AVvLMv6gmMNam3uVAjZpfkcJCwDwnZn6z3wXmqPV".into(),
-        )
-        .unwrap();
-        new.add_key_pair(public_key.clone());
-
-        let json_diff = serde_json::to_string(&Diff::serializable(&old, &new)).unwrap();
-        let mut deserializer = serde_json::Deserializer::from_str(&json_diff);
-        Apply::apply(&mut deserializer, &mut old).unwrap();
-
-        assert_eq!(old.to_string(), new.to_string());
-    }
-}
